changed adminNumber to 7041196959 in opeateOnMsg() :           changes datatype of data variable to unsigned short int from int.           reduced the response length of status msg,
         changed the BAL check, so that AT+CUSD=1 is not needed to send each time,

reduced string responses at various places,
changed the signature of stringContains, to accept byte arguments instead of int arguements for start and stop index.
added wait in sending SMS after sending the Ctrl+Z character i.e. (char)26
changes in update():
        send AT+CUSD=0 command after bal msg is sent.

+++ Motor/SIM.h
@@ -78,7 +78,7 @@ class SIM
     bool sendBlockingATCommand(String);
     String readString();
     bool matchString(String, String);
-    bool stringContains(String &sstr, String mstr, int sstart, int sstop);
+    bool stringContains(String &sstr, String mstr, byte sstart, byte sstop);
     bool isRinging(String);
     bool isDTMF(String &str);
     bool isCut(String);
